+++ packages/sdk/ts_js/src/barrits_lib/logic/algorithms/timeseries/types.ts
@@ -5,7 +5,9 @@
  * @template Value [EN] Type of the data value. [ES] Tipo del valor de los datos.
  */
 export type TimeSeriesPoint<Value = number> = {
+  /** [EN] Timestamp. [ES] Marca de tiempo. */
   readonly timestamp: number;
+  /** [EN] Value. [ES] Valor. */
   readonly value: Value;
 };
 
@@ -14,7 +16,10 @@ export type TimeSeriesPoint<Value = number> = {
  * [ES] Representa una cubeta temporal que contiene múltiples puntos de series temporales.
  */
 export type TimeBucket<Value = number> = {
+  /** [EN] Bucket start. [ES] Bucket inicio. */
   readonly bucketStart: number;
+  /** [EN] Bucket end. [ES] Bucket fin. */
   readonly bucketEnd: number;
+  /** [EN] Points. [ES] Puntos. */
   readonly points: TimeSeriesPoint<Value>[];
 };